+++ src/Products/PloneMeeting/browser/itemattendee.py
@@ -207,13 +207,15 @@ class ByeByeAttendeeForm(BaseAttendeeForm):
             if self.context.show_votes():
                 voters = item_to_update.get_item_voters()
                 if self.person_uid in voters:
-                    # secret
-                    if item_to_update.get_votes_are_secret():
-                        # is there place to remove a voter?
-                        len_voters = len(voters)
-                        all_item_votes = item_to_update.get_item_votes()
-                        i = 0
-                        for item_vote in all_item_votes:
+                    all_item_votes = item_to_update.get_item_votes(
+                        ignored_vote_values=[NOT_ENCODED_VOTE_VALUE])
+                    i = 0
+                    # used when vote is secret
+                    len_voters = len(voters)
+                    for item_vote in all_item_votes:
+                        # secret
+                        if item_to_update.get_vote_is_secret(vote_number=i):
+                            # every voters voted?
                             encoded_votes_count = item_to_update.getVoteCount(
                                 vote_value='any_voted', vote_number=i)
                             if len_voters == encoded_votes_count:
@@ -227,23 +229,19 @@ class ByeByeAttendeeForm(BaseAttendeeForm):
                                     type='warning',
                                     request=self.request)
                                 error = True
-                    # public
-                    else:
-                        all_item_votes = item_to_update.get_item_votes(
-                            ignored_vote_values=[NOT_ENCODED_VOTE_VALUE])
-                        hp_uid_in_voters = bool([item_vote for item_vote in all_item_votes
-                                                 if self.person_uid in item_vote['voters']])
-                        if hp_uid_in_voters:
-                            api.portal.show_message(
-                                _("Can not set ${not_present_type} "
-                                  "a person that voted on an item!",
-                                  mapping={
-                                      'not_present_type':
-                                          _('item_not_present_type_{0}'.format(
-                                            self.not_present_type))}),
-                                type='warning',
-                                request=self.request)
-                            error = True
+                        # public
+                        else:
+                            if self.person_uid in item_vote['voters']:
+                                api.portal.show_message(
+                                    _("Can not set ${not_present_type} "
+                                      "a person that voted on an item!",
+                                      mapping={
+                                          'not_present_type':
+                                              _('item_not_present_type_{0}'.format(
+                                                self.not_present_type))}),
+                                    type='warning',
+                                    request=self.request)
+                                error = True
 
             if error:
                 if item_to_update != self.context:
@@ -271,8 +269,7 @@ class ByeByeAttendeeForm(BaseAttendeeForm):
         # user will first have to select another signatory on meeting or item
         # return a portal_message if trying to set absent and item that is
         # already excused (and the other way round)
-        error = self._mayByeByeAttendeePrecondition(items_to_update)
-        if error:
+        if self._mayByeByeAttendeePrecondition(items_to_update):
             self._finished = True
             return