+++ src/Products/PloneMeeting/browser/itemvotes.py
@@ -292,7 +292,8 @@ def is_vote_updatable_for(context, item_to_update):
        if using same pollType and same voters and not using several or linked votes."""
     res = False
     if context == item_to_update or \
-       (context.getPollType() == item_to_update.getPollType() and
+       (context.get_item_votes(vote_number=0).get('poll_type', context.getPollType()) ==
+        item_to_update.get_item_votes(vote_number=0).get('poll_type', item_to_update.getPollType()) and
         context.get_item_voters() == item_to_update.get_item_voters() and
             len(item_to_update.get_item_votes()) < 2):
         res = True
@@ -350,10 +351,12 @@ class EncodeVotesForm(BaseAttendeeForm):
                 if wdt.__name__ == 'voter_uid':
                     wdt.mode = HIDDEN_MODE
         # disable apply_until_item_number when using several votes
-        if len(self.context.get_item_votes()) > 1 or self.request.get('vote_number', 0) > 0:
+        # or if poll_type was redefined
+        if _should_disable_apply_until_item_number(self.context):
             apply_until_item_number = self.widgets['apply_until_item_number']
             apply_until_item_number.disabled = "disabled"
-            apply_until_item_number.title = _(u"Not available when using several votes on same item.")
+            apply_until_item_number.title = _(
+                u"Not available when using several votes on same item.")
 
     def _doApply(self):
         """ """
@@ -370,6 +373,10 @@ class EncodeVotesForm(BaseAttendeeForm):
         data['voters'] = PersistentMapping()
         for vote in self.votes:
             data['voters'][vote['voter_uid']] = vote['vote_value']
+        # add poll_type in case it was redefined
+        vote_infos = self.context.get_item_votes(vote_number=self.vote_number)
+        if 'poll_type' in vote_infos:
+            data['poll_type'] = vote_infos['poll_type']
 
         items_to_update = _itemsToUpdate(
             from_item_number=self.context.getItemNumber(relativeTo='meeting'),
@@ -542,6 +549,17 @@ class IEncodeSecretVotes(Interface):
             raise Invalid(msg)
 
 
+def _should_disable_apply_until_item_number(context):
+    """Is it possible to use the "apply_until_item_number" field?"""
+    vote_number = context.REQUEST.get('vote_number', 0)
+    if len(context.get_item_votes()) > 1 or \
+       vote_number > 0 or \
+       context.get_item_votes(vote_number).get('poll_type', None) != \
+       context.getPollType():
+        return True
+    return False
+
+
 class EncodeSecretVotesForm(BaseAttendeeForm):
     """ """
 
@@ -569,7 +587,7 @@ class EncodeSecretVotesForm(BaseAttendeeForm):
                 elif wdt.__name__ == 'vote_value' and wdt.value:
                     wdt.klass += " {0}".format(wdt.context['vote_value'])
         # disable apply_until_item_number when using several votes
-        if len(self.context.get_item_votes()) > 1 or self.request.get('vote_number', 0) > 0:
+        if _should_disable_apply_until_item_number(self.context):
             apply_until_item_number = self.widgets['apply_until_item_number']
             apply_until_item_number.disabled = "disabled"
             apply_until_item_number.title = _(u"Not available when using several votes on same item.")
@@ -665,7 +683,7 @@ class ItemDeleteVoteView(BrowserView):
             assert self.context._voteIsDeletable(vote_number)
 
             vote_to_delete = item_votes[vote_number]
-            if self.context.get_votes_are_secret():
+            if self.context.get_vote_is_secret(vote_number):
                 originnal_vote_keys = [str(vote_count) for vote_value, vote_count
                                        in vote_to_delete['votes'].items()]
                 originnal_vote_keys = "__".join(originnal_vote_keys)